+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ### Added
 
 - Add include condition to inherited rules.
+- Modify use/loader fields to get correct module path.
 
 ## [0.1.0] - 2026-03-23